+++ src/api/api-utils.js
@@ -1,6 +1,6 @@
 const getDate = (req) => {
   let date = new Date();
-  const queryDate = req.query.date || req.query.fromDate;
+  const queryDate = req.query.date;
   if (queryDate) {
     date = new Date(queryDate) || date;
   }
@@ -30,6 +30,7 @@ const createCsvElements = (data) => {
 const getHourlyDates = (date) => {
   const todayHours = [];
   date.setUTCHours(0);
+  todayHours.push(new Date(date));
   while (date.getUTCHours() < 23) {
     date.setUTCHours(date.getUTCHours() + 1);
     todayHours.push(new Date(date));
@@ -38,10 +39,23 @@ const getHourlyDates = (date) => {
   return todayHours;
 };
 
+const getHoursBetweenDates = (dateFrom, dateTo) => {
+  const hours = [];
+  const from = new Date(dateFrom);
+  const to = new Date(dateTo);
+  hours.push(new Date(from));
+  while (from < to) {
+    from.setUTCHours(from.getUTCHours() + 1);
+    hours.push(new Date(from));
+  }
+  return hours;
+};
+
 export {
   getDate,
   getOverlapped,
   createCsvElements,
   getGroupField,
   getHourlyDates,
+  getHoursBetweenDates,
 }; // eslint-disable-line import/prefer-default-export