+++ src/api/middleware/auth-check-middleware.js
@@ -5,4 +5,5 @@ import jwt from 'express-jwt';
 export default ({ secret, clientId }) => () => jwt({
   secret,
   audience: clientId,
+  algorithms: ['RS256', 'HS256'],
 });