+++ src/events/mapper/toHandlers/s6fresnel/S6InfoToDevice.js
@@ -1,6 +1,12 @@
-const makeCommands = (location, deviceId) => Object.assign({}, {
-  power: `mqtt:building/${location}/devices/${deviceId}/power`,
-});
+const makeCommands = (topic, gateway, location, deviceId) => {
+  const topicParts = (topic || '').split('/');
+  let root = gateway;
+  // Retro-compatibilty check to manage device using old constant "building" as topic root
+  if (topicParts.length > 0 && topicParts[0] === 'building') {
+    root = topicParts[0];
+  }
+  return Object.assign({}, { power: `mqtt:${root}/${location}/devices/${deviceId}/power` });
+};
 
 const S6InfoToDevice = e => ({
   deviceId: e.Payload.deviceId || '00:00:00:00:00:00',
@@ -13,7 +19,7 @@ const S6InfoToDevice = e => ({
     deviceId: e.Payload.deviceId || '00:00:00:00:00:00',
     group: e.Payload.group,
     features: e.Payload.features || [],
-    commands: makeCommands(e.Payload.group, e.Payload.deviceId),
+    commands: makeCommands(e.Payload.topic, e.GatewayId, e.Payload.group, e.Payload.deviceId),
     created: new Date(),
   },
 });